improved "auto_complete_visible" detection
In a recent patch, lispindent was set to not be used
if the auto complete is visible. This to make it possible
to commit on enter.

However, if the setting
"auto_complete_commit_on_tab" is set to true, enter
should always insert a new line - even if auto complete
is visible - and it should be indented.

The problem is that if the auto complete is visible, lispindent
will NEVER be used. This patch fixes that by also checking
if the option mentioned above is set to true.